refactor: move util functions to their own files
